Fix review findings
